first commit & explanation of script
It will move files in a folder into subfolders. It works with any type of file. The number 11 in the script represents the number of files you want to be moved to each subfolder. This number can be changed. If the files should be divided into batches of 5 and moved to subfolders then you would replace 11 with 5.